Still more changes in master??! I just committed it all, and changed NOTHING?!

+++ api/src/database/CourseRegistrationHelper.ts
@@ -1,13 +1,15 @@
-import {query, extract, map, one, toBin} from "./HelperDB";
+const HH = require("./HelperHelper");
 
 import {CourseRegistration, convertCourseReg} from '../../../models/CourseRegistration';
-import {RolePermissionDB} from './RolePermissionDB'
+import RolePermissionHelper from './RolePermissionsHelper'
 /**
  * courseID, userID, role, permission
  * @Author Rens Leendertz
  */
+const {query, extract, map, one} = HH;
+export default class CourseRegistrationHelper {
 
- export class CourseRegistrationDB {
+	static toBin = RolePermissionHelper.toBin;
 
 	/**
 	 * return all entries in this table, with permissions set correctly
@@ -71,9 +73,9 @@ import {RolePermissionDB} from './RolePermissionDB'
 			role,
 			permission = 0
 		} = entry;
-		return query(`INSERT INTO "CourseRegistration" 
+		return query(`INSERT INTO \"CourseRegistration\" 
 			VALUES ($1,$2,$3,$4) 
-			RETURNING *`, [courseID, userID, role, toBin(permission)])
+			RETURNING *`, [courseID, userID, role, CourseRegistrationHelper.toBin(permission)])
 		.then(extract).then(map(convertCourseReg)).then(one)
 	}
 
@@ -88,10 +90,11 @@ import {RolePermissionDB} from './RolePermissionDB'
 			userID,
 			role
 		} = entry;
-		return query(`UPDATE "CourseRegistration" SET 
+		return query(`UPDATE \"CourseRegistration\" SET 
 			courseRole=COALESCE($3, courseRole)
 			WHERE courseID=$1 AND userID=$2
-			RETURNING *`, [courseID, userID, role])
+			RETURNING *
+			`, [courseID, userID, role])
 		.then(extract).then(map(convertCourseReg)).then(one)
 	}
 
@@ -106,10 +109,11 @@ import {RolePermissionDB} from './RolePermissionDB'
 			userID,
 			permission
 		} = entry;
-		return query(`UPDATE "CourseRegistration" SET 
+		return query(`UPDATE \"CourseRegistration\" SET 
 			permission=permission | $3
 			WHERE courseID=$1 AND userID=$2
-			RETURNING *`, [courseID, userID, toBin(permission)])
+			RETURNING *
+			`, [courseID, userID, CourseRegistrationHelper.toBin(permission)])
 		.then(extract).then(map(convertCourseReg)).then(one)
 	}
 
@@ -125,10 +129,11 @@ import {RolePermissionDB} from './RolePermissionDB'
 			userID,
 			permission
 		} = entry;
-		return query(`UPDATE "CourseRegistration" SET 
+		return query(`UPDATE \"CourseRegistration\" SET 
 			permission=permission & ~($3::bit(40))
 			WHERE courseID=$1 AND userID=$2
-			RETURNING *`, [courseID, userID, toBin(permission)])
+			RETURNING *
+			`, [courseID, userID, CourseRegistrationHelper.toBin(permission)])
 		.then(extract).then(map(convertCourseReg)).then(one)
 	}
 
@@ -141,9 +146,7 @@ import {RolePermissionDB} from './RolePermissionDB'
 			courseID,
 			userID
 		} = entry;
-		return query(`DELETE FROM "CourseRegistration" 
-			WHERE courseID=$1 AND userID=$2 
-			RETURNING *`, [courseID, userID])
+		return query("DELETE FROM \"CourseRegistration\" WHERE courseID=$1 AND userID=$2 RETURNING *", [courseID, userID])
 		.then(extract).then(map(convertCourseReg)).then(one)
 	}
 }
